scripts/requirements.txt
requests
beautifulsoup4
lxml
pyyaml
black
pytz
